Add support for multiple SD cards on one host interface.
Needs external logic to multiplex the clock signal,
which is controlled by a GPIO output pin.

Only the selected SD card receives the clock,
while the other SD card's clock is forced to LOW.
DATA and CMD lines can be connected to both SD cards.

On warm-start UBOOT should use a sequence of
CMD12 and CMD7 with RCA=0 to force any active SD card
to stop transmission and release the DATA bus.

Instantiate this in the device tree as:

 &mmc {
  select-gpio = <&portd 2 0>;
  cd-gpios = <&portb 28 1>,
             <&porta 18 1>;
 };

Signed-off-by: Bernd Edlinger <bernd.edlinger@hotmail.de>