socfpga: Fix wrong result in determine_rate hook
This fixes a wrong baud rate computation, that results
in garbage output of the serial console.

For socfpga devices, clocks are divided from the base clock
with a fixed or otherwise pre-determined divisor, but the function
clk_hw_determine_rate_no_reparent just returns the parent's
clock frequency unmodified which leads to wrong results.

Fixed that by returning the same value as the recalc_rate hook.

Fixes: 9607beb917df ("clk: socfpga: gate: Add a determine_rate hook")
Signed-off-by: Bernd Edlinger <bernd.edlinger@hotmail.de>